Move sections to their own files
