+++ src/components/challenge/ChallengeSection.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const ChallengeSection = () => {
+  return (
+      <div className="container-fluid">
+        <div className="row">
+          <h2 className="challenge-section-title">Challenge</h2>
+          <p className="challenge-section-desc">Lorem Ipsum</p>
+        </div>
+      </div>
+  )
+}
+
+export default ChallengeSection;
