+++ src/components/challenge/CheckInSection.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const CheckInSection= () => {
+  return (
+      <div className="container-fluid">
+        <div className="row">
+          <h2 className="challenge-section-title">Check-In</h2>
+          <p className="challenge-section-desc">Lorem Ipsum</p>
+        </div>
+      </div>
+  )
+}
+
+export default CheckInSection;
