+++ src/components/challenge/PrizesSection.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const PrizesSection= () => {
+  return (
+      <div className="container-fluid">
+        <div className="row">
+          <h2 className="challenge-section-title">Prizes</h2>
+          <p className="challenge-section-desc">Lorem Ipsum</p>
+        </div>
+      </div>
+  )
+}
+
+export default PrizesSection;
